added services folder and moved api fetches to LoginPage.servie.jsx from LoginPage.jsx
